리뷰 반영
1. 불필요한 line break 제거
2. useState, handleClick 내용 수정
3. export default 위치 수정
4. index.jsx, App.jsx 분리
