+++ .eslintrc.js
@@ -45,7 +45,6 @@ module.exports = {
   'object-curly-spacing': ['error', 'always'],
   'key-spacing': ['error', { mode: 'strict' }],
   'arrow-spacing': ['error', { before: true, after: true }],
-
   'react/prop-types': 'off',
   'react/jsx-indent': ['error', 1],
   'react/jsx-indent-props': ['error', 1],
