+++ src/index.jsx
@@ -1,28 +1,6 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
-import CounterPage from './pages/CounterPage';
-
-function App() {
- const [state, setState] = useState({
-  count: 0,
- });
-
- const { count } = state;
-
- function handleClick(size) {
-  setState({
-   count: count + size,
-  });
- }
-
- return (
-  <CounterPage
-   count={count}
-   onClick={() => handleClick(1)}
-   onClickNumber={(number) => handleClick(number)}
-  />
- );
-}
+import App from './App';
 
 ReactDOM.render(
  <App />,
